setting/package-structure/FLOW-0 패기지 구조 mvc 패턴 및 기능 기반으로 개편, 추후 각각의 파일 수정 필요
