--- package.json
@@ -1,82 +1,0 @@
-{
-  "name": "chatter",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "webpack serve --mode=development --hot",
-    "build": "webpack --mode=production",
-    "start": "npx serve -s dist",
-    "lint": "eslint .",
-    "format": "prettier --check ./src",
-    "format.fix": "prettier --write ./src"
-  },
-  "dependencies": {
-    "@reduxjs/toolkit": "^2.6.1",
-    "@sentry/react": "^9.10.0",
-    "@sentry/tracing": "^7.120.3",
-    "@stomp/stompjs": "^7.1.0",
-    "@tanstack/react-query": "^5.69.0",
-    "axios": "^1.7.9",
-    "clsx": "^2.1.1",
-    "firebase": "^11.5.0",
-    "framer-motion": "^12.6.2",
-    "idb": "^8.0.2",
-    "radix-ui": "^1.1.3",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hook-form": "^7.54.2",
-    "react-redux": "^9.2.0",
-    "react-router-dom": "^7.0.2",
-    "reconnecting-websocket": "^4.4.0",
-    "simple-peer": "^9.11.1",
-    "sockjs-client": "^1.6.1",
-    "uuid": "^11.0.4",
-    "zod": "^3.24.2"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.26.8",
-    "@babel/preset-react": "^7.26.3",
-    "@babel/preset-typescript": "^7.26.0",
-    "@eslint/js": "^9.16.0",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.2.0",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.10.2",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
-    "@types/react-redux": "^7.1.34",
-    "@typescript-eslint/eslint-plugin": "^8.28.0",
-    "@typescript-eslint/parser": "^8.28.0",
-    "@vitejs/plugin-react": "^4.3.4",
-    "autoprefixer": "^10.4.20",
-    "babel-loader": "^9.2.1",
-    "clean-webpack-plugin": "^4.0.0",
-    "css-loader": "^7.1.2",
-    "cypress": "^14.2.1",
-    "eslint": "^9.23.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-react": "^7.37.4",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "eslint-plugin-tailwindcss": "^3.18.0",
-    "globals": "^15.13.0",
-    "html-webpack-plugin": "^5.6.3",
-    "husky": "^9.1.7",
-    "jest": "^29.7.0",
-    "lint-staged": "^15.5.0",
-    "postcss": "^8.4.49",
-    "postcss-loader": "^8.1.1",
-    "prettier": "3.4.2",
-    "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.17",
-    "typescript": "~5.6.2",
-    "typescript-eslint": "^8.18.0",
-    "webpack": "^5.97.1",
-    "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^5.2.0",
-    "workbox-webpack-plugin": "^7.3.0"
-  }
-}
