+++ src/lib/utils/auth/stringValidations.ts
