--- src/pages/Welcome.tsx
@@ -1,11 +1,0 @@
-import { useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
-
-export default function Welcome() {
-  const navigate = useNavigate()
-  useEffect(() => {
-    navigate('/login')
-  }, [])
-
-  return <div />
-}
