+++ src/view/pages/home/components/DirectBar.tsx
