+++ src/view/pages/home/components/Icon.tsx
