+++ src/view/pages/home/components/Interactions.tsx
