+++ src/view/pages/home/components/NavigationCard.tsx
